+++ src/Cocoar.Configuration/Providers/LocalStorageProvider/LocalStorageProviderQueryOptions.cs
@@ -1,0 +1,8 @@
+using Cocoar.Configuration.Providers.Abstractions;
+
+namespace Cocoar.Configuration.Providers;
+
+public sealed class LocalStorageProviderQueryOptions : IProviderQuery
+{
+    public static readonly LocalStorageProviderQueryOptions Default = new();
+}